(no-tn-check) Fix for polymorphic by-ref objects; streaming
Fix various problems associate with by-ref objects, in particular
polymorphic by-ref parameters were not properly handled, meaning
that no polymorphic "wrapper" was getting created when the
actual parameter was a "monomorphic" type.

The debugger console was not displaying by-ref components
properly.  The debugger console was also crashing in some
cases when given uninitialized objects.  Now it is a bit
more fault tolerant.

We have implemented the beginning of byte streaming for files,
as part of the interpreter.  This will eventually be accompanied
by a set of library modules for byte, value, and object streaming.