(no-tn-check) Improve Unicode support
We now handle the \#nnnn# syntax in both string and character
literals as a way to include characters with the given
Unicode code.  We now handle Unicode characters with codes
above 0xFFFF more completely.

This includes a fix for a bug which was forcing all strings
to be "uniquified" rather than being managed dynamically.
Now we only uniquify strings at compile time; at run time
strings are handled more dynamically with no attempt to
look up in a hash table.