(no-tn-check) Allow use of "limited" to indicate not assignable
When declaring an abstract interface, it is important to know
whether or not it is "assignable" because an extension
of an assignable module should itself be assignable.
We adopt the Ada term "limited" for an interface that is *not*
assignable.

Also update .gitignore to ignore a bit more stuff, and add
a source position to a call on run_time_type_info in a case
where we might produce an error message.