(no-tn-check) Work on byte, value, and object streaming
Beginning to implement byte, value, and object streaming
which will be important for distributed communications as
well.  Have switched to using mode "in out" for
Exec_Context in most operations, to minimize need to
create new execution contexts when invoking built-in
routines.

Switched to using "ref interface" rather than "limited interface"
to indicate an interface that is not assignable, and is
expected to have "ref" components.

Miscellaneous other bug fixes.