(no-tn-check) Add Strict_Compare operations for full ordering
The Ordered interface now requires a Strict_Compare operation
which never returns #unordered when comparing two items.
Strict_Compare is used in Ordered_Sets and Ordered_Maps to allow
keying on types where "=?" is a partial ordering.

We have now added the Strict_Compare to all types in aaa.psi
that have a full ordering defined.

We also fixed a problem where two operation were not being
recognized as homonyms, even though all the parameter/result
types were the same, due to analysis ordering issues
(in particualar for the type Ordered::Full_Ordering).
We now allow them to match if they refer to the same symbol,
even if the actual "base" Type info has not been filled in.