(no-tn-check)  Add notion of "others" branching node
An "others" branching node is used when there is a nested
pattern match labeled with an "others".  This is treated
specially to ensure none of the end states reachable from this
others are considered ambiguous with any end states preceding
the "others" branch.  It is truly an "others" rather than an "any".
This normally doesn't matter because of our "superset" rule,
but in this case, the end states reachable from the others
might not all be supersets of states preceding them if we
interpret the "others" as an "any value."