Use "Interfaces.Work_Stealing"; Fix deque expansion race-condition.
We now use "Interfaces.Work_Stealing" rather than "Interfaces_Work_Stealing"
because we are passing the -gnatg flag which allows us to define
packages whose full name starts with "Ada.", "System.", or
"Interfaces.".

Use term "Capacity" rather than "Modulus" as it seems clearer
that it represents the capacity of the deque.

Rewrite the code that handles deque expansion to avoid a race
condition where a "steal" is underway while "push" doubles the
size of the deque.  Before we were just doing an Unchecked_Deallocation
of the old deque without worrying about this possibility.  Now we add
the Capacity to Top using Atomic_Compare_And_Exchange and then
add the same amount to Bottom.  This will cause any ongoing "steal"
to fail if it starts before the expansion and tries to use an element
from a deallocated deque.

Shrink the initial capacity of a deque to 16 from 256, as we generally
start stealing from a deque before it gets very big.