Reclaim storage for Work-Stealing TCB
Valgrind, as reported by Egil Høvik, found a storage
leak for the allocator "new WS_Data'(...)".  We now
reclaim that storage as part of Finish_Sub_LWT.  This
involves changing the mode of the TCB to be "in out" and
calling an instance of Unchecked_Deallocation as the
last step.