When finishing type descriptors, reset the error count before each one.
By resetting the error count when we hit, for example, some kind
of mismatch while finishing a type descriptor, we can proceed
to finish other type descriptors, and find other errors.
Ultimately, we set the error count to the total number of
errors, to prevent the program from being executed, and so
the error reporting has the correct count.