More work on Phiddle (HDL in ParaSail) and sequenced containers
We have added a new kind of set and map, namely a "sequenced"
set and map, which retain the original insertion order.  This is used
in "phiddle", the prototype HDL in ParaSail.  Phiddle has been
augmented so that it is beginning to have the features needed to
actually define a circuit.