Provide a default for "After" in Sys.Set_Value.
Provide a default of No_Delay for Sys.Set_Value so it
can be used instead of Signal.Set_Value or Port.Set_Value
without having to add an explicit No_Delay.  Note that
the Sys.Set_Value will always put the update onto the
queue, even if it has a zero delay.  It will then be
pulled back off the queue when Do_Processing is called,
providing the effective delayed effect of an assignment
as present in VHDL for signals.