Fix problem with polymorphic constant initialized from existing const
The code-gen phase was dying when trying to simply re-use a (monomorphic)
constant, when in fact we wanted a corresponding polymorphic constant,
as in the following example:

   const Child_Inst : Child := ...
   const Parent_Inst : Parent+ := Child_Inst;

where Child is an extension of the Parent type.

We now treat this situation as though it were a completely new
computation, and we execute a little program to create the
polymorphic constant.