Enhance xvector example; display comp names when dumping an obj
Enhance the xvector example to test some new capabilities
including the new "turbofish" syntax for instantiating
a module in the middle of an expression, such as:

   var x := Vector::<Integer>::[1, 2, 3, 4];

where "Vector::<" is used instead of simply "Vector<" when
in an expression context where "<" would be ambiguous (could
be a "less than" operator).

Also, now that Type_Descriptor Component_Info's have a Decl
component, use that to find the name of the component, so
it can be displayed when displaying the components of a type
descriptor.