Implement an "early exit" capability in parallel loops
There is now a new routine "Par_Range_Loop_With_Early_Exit"
and a new function "Early_Exit" in LWT.Parallelism.
The Loop_Body parameter to Par_Range_Loop_With_Early_Exit has
an additional PID parameter, which can be passed to the
Early_Exit function to attempt to cancel a thread group.  If it
returns True, the calling thread was the first to attempt to
cancel the group, and it should finish the work to complete the
loop.  If Early_Exit returns False, then some other thread already
canceled the group, and the calling thread should simply return and
do no more processing.

We added a PID parameter to the Loop_Body parametr of
Generic_Par_Iterator_Loop as well so early exits are always
possible.  This avoided having to have two generic procedures
differing only in the parameters to the passed-in Loop_Body parameter.

In any case, the PID parameter may be ignored if there is no
early exit from the loop body.