Move Generic_Par_Iterator_Loop inside parallel interface package
To simplify the job of the compiler, we have eliminated the need
for a separate instantiation of Generic_Par_Iterator_Loop, by
moving it into the Parallel_Iterator_Interfaces package.

Ultimately this procedure should be in the private part of
the generic package, or at least in some place that it won't "pollute"
the namespace of the Iterator_Interfaces package.