Add two Ada2022 parallel test programs; update do_atags
n_queens.adb takes a count of number of rows in the chessboard,
and produces a list of the various solutions to the N-Queens problem.
sorting.ads/adb and test_sorting.adb provide a simple parallel quicksort,
which takes a count of the size of the integer array to be sorted.
It has been verified that [test_]sorting.ads/adb compile and run when
the "parallel do" and "and" keywords are replaced with comment lines.
N_Queens is based very closely on a program that works sequentially.

do_atags.csh has been updated to include lwt in the "tags" database.